Implement configurable tools list
Specify tools in config

Send info about all tools to UI

Change machine configuration format

+++ ModbusSender/IModbusSender.cs
@@ -5,6 +5,6 @@ namespace ModbusSender
     public interface IModbusSender
     {
         void Init();
-        void WriteRegister(UInt16 devid, UInt16 index, UInt16 value);
+        void WriteRegister(int devid, UInt16 index, UInt16 value);
     }
 }
